backend/requirements.txt
fastapi
uvicorn
gitpython
requests
python-dotenv
slowapi
